Improve support for reconnection of the graph on changing sources.
Adds in support for including a custom filter in the graph construction
phase and maintaining it during reconnection.
This resolves an issue where an older video source interleaved with the
newly selected source after changing sources. Sometimes the webcam would
interleave with a video file source if a file source was selected after
initializing with the webcam.